Remove EDH campaigns api methods & tests
