+++ source/api/campaigns/index.js
@@ -1,34 +1,39 @@
-import { isJustGiving } from '../../utils/client'
+import get from 'lodash/get'
+import { servicesAPI } from '../../utils/client'
+import { baseUrl, imageUrl, parseText } from '../../utils/justgiving'
 import { required } from '../../utils/params'
 
-import {
-  fetchCampaigns as fetchJGCampaigns,
-  fetchCampaign as fetchJGCampaign,
-  fetchCampaignGroups as fetchJGCampaignGroups,
-  deserializeCampaign as deserializeJGCampaign
-} from './justgiving'
-
-import {
-  fetchCampaigns as fetchEDHCampaigns,
-  fetchCampaign as fetchEDHCampaign,
-  fetchCampaignGroups as fetchEDHCampaignGroups,
-  deserializeCampaign as deserializeEDHCampaign
-} from './everydayhero'
-
-export const fetchCampaigns = (params = required()) => {
-  return isJustGiving() ? fetchJGCampaigns(params) : fetchEDHCampaigns(params)
+export const fetchCampaigns = ({ ids = required() }) => {
+  return Promise.resolve()
+    .then(() => (Array.isArray(ids) ? ids : ids.split(',')))
+    .then(ids => Promise.all(ids.map(fetchCampaign)))
 }
 
-export const fetchCampaign = (id = required()) => {
-  return isJustGiving() ? fetchJGCampaign(id) : fetchEDHCampaign(id)
-}
+export const fetchCampaign = (id = required()) =>
+  servicesAPI
+    .get(`/v1/justgiving/campaigns/${id}`)
+    .then(response => response.data)
 
-export const fetchCampaignGroups = (id = required()) => {
-  return isJustGiving() ? fetchJGCampaignGroups(id) : fetchEDHCampaignGroups(id)
-}
-
-export const deserializeCampaign = campaign => {
-  return isJustGiving()
-    ? deserializeJGCampaign(campaign)
-    : deserializeEDHCampaign(campaign)
-}
+export const deserializeCampaign = campaign => ({
+  charities: campaign.charities,
+  donateUrl: [
+    baseUrl('link'),
+    'v1/campaign/donate/campaignGuid',
+    campaign.campaignGuid
+  ].join('/'),
+  eventId: campaign.eventId,
+  getStartedUrl: `${baseUrl()}/fundraising-page/creation?campaignGuid=${
+    campaign.campaignGuid
+  }`,
+  id: campaign.campaignGuid,
+  name: campaign.title,
+  image: imageUrl(get(campaign, 'heroImage.imageName')),
+  raised: campaign.donationSummary.totalAmount,
+  raisedOffline: campaign.donationSummary.offlineAmount,
+  slug: campaign.shortName,
+  story: parseText(campaign.story),
+  summary: campaign.summary,
+  target: campaign.targetAmount,
+  totalDonations: campaign.donationSummary.totalNumberOfDonations,
+  url: `${baseUrl()}/campaign/${campaign.shortName}`
+})
