+++ source/api/campaigns/readme.md
@@ -4,7 +4,6 @@ Helpers related to fetching campaigns
 
 - [fetchCampaigns](#fetchcampaigns)
 - [fetchCampaign](#fetchcampaign)
-- [fetchCampaignGroups](#fetchcampaigngroups)
 - [deserializeCampaign](#deserializecampaign)
 
 
@@ -31,7 +30,7 @@ A pending promise that will either resolve to:
 import { fetchCampaigns } from 'supporticon/api/campaigns'
 
 fetchCampaigns({
-  charity: 'au-28'
+  ids: ['96e2266e-2fa2-4109-a2b6-c017b79011bd']
 })
 ```
 
@@ -57,32 +56,7 @@ A pending promise that will either resolve to:
 ```javascript
 import { fetchCampaign } from 'supporticon/api/campaigns'
 
-fetchCampaign('au-6839')
-```
-
-## `fetchCampaignGroups`
-
-**Purpose**
-
-Fetch the groups for a campaign.
-
-**Params**
-
-- `id` The id of the campaign
-
-**Returns**
-
-A pending promise that will either resolve to:
-
-- Success: the data returned from the request
-- Failure: the error encountered
-
-**Example**
-
-```javascript
-import { fetchCampaignGroups } from 'supporticon/api/campaigns'
-
-fetchCampaignGroups('au-6839')
+fetchCampaign('96e2266e-2fa2-4109-a2b6-c017b79011bd')
 ```
 
 ## `deserializeCampaign`
